Move Unit test project to the tests subfolder
