--- E.Deezer.Tests/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.2.1" targetFramework="net45" />
-  <package id="Moq" version="4.8.0" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net462" />
-  <package id="NUnit" version="3.11.0" targetFramework="net45" />
-  <package id="NUnit3TestAdapter" version="3.12.0" targetFramework="net45" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net45" />
-  <package id="System.Threading.Tasks.Extensions" version="4.4.0" targetFramework="net45" />
-</packages>
