+++ E.Deezer.sln
@@ -8,11 +8,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{84ABD1
 		.nuget\NuGet.Config = .nuget\NuGet.Config
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "E.Deezer.Tests", "E.Deezer.Tests\E.Deezer.Tests.csproj", "{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "E.Deezer", "E.Deezer\E.Deezer.csproj", "{7DC0FD2E-3A4F-49A2-94F1-E54D51835948}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{697A9BE3-3B02-4CA7-B0B0-A6747550A8ED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Unit", "tests\Unit\Unit.csproj", "{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Integration", "tests\Integration\Integration.csproj", "{DAB15430-0472-40D5-9FDB-66BE8FAA3596}"
 EndProject
@@ -24,14 +24,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7DC0FD2E-3A4F-49A2-94F1-E54D51835948}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7DC0FD2E-3A4F-49A2-94F1-E54D51835948}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7DC0FD2E-3A4F-49A2-94F1-E54D51835948}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7DC0FD2E-3A4F-49A2-94F1-E54D51835948}.Release|Any CPU.Build.0 = Release|Any CPU
+		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DAB15430-0472-40D5-9FDB-66BE8FAA3596}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DAB15430-0472-40D5-9FDB-66BE8FAA3596}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DAB15430-0472-40D5-9FDB-66BE8FAA3596}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -45,6 +45,7 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{728BFDB3-DCA8-464C-91A5-B5AC2E873ADA} = {697A9BE3-3B02-4CA7-B0B0-A6747550A8ED}
 		{DAB15430-0472-40D5-9FDB-66BE8FAA3596} = {697A9BE3-3B02-4CA7-B0B0-A6747550A8ED}
 		{7F3A3B85-D101-48A8-B5FD-273B487C32F2} = {697A9BE3-3B02-4CA7-B0B0-A6747550A8ED}
 	EndGlobalSection
