+++ tests/Unit/Api/Internal/PermissionsTests.cs
@@ -1,7 +1,7 @@
 ﻿using E.Deezer.Api;
 using NUnit.Framework;
 
-namespace E.Deezer.Tests.Api.Internal
+namespace E.Deezer.Tests.Unit.Api.Internal
 {
     [TestFixture]
     class PermissionsTests
