+++ tests/Unit/Exception/DeezerExceptionTests.cs
@@ -2,7 +2,7 @@
 
 using NUnit.Framework;
 
-namespace E.Deezer.Tests.Exception
+namespace E.Deezer.Tests.Unit.Exception
 {
     [TestFixture]
     [Parallelizable(ParallelScope.All)]
