+++ tests/Unit/Resources/Api/Album.json
