+++ tests/Unit/Resources/Api/Internal/ObjectWithImageDefault.json
