+++ tests/Unit/Resources/Api/Internal/ObjectWithImageMissingPicture.json
