+++ tests/Unit/Resources/Api/Internal/ObjectWithImageOnlyCovers.json
