+++ tests/Unit/Str.cs
